# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "✨Reverse-engineered API for Gemini web app. It can be used as a genuine API key from OpenAI, Gemini, and Claude.",
        "title": "Gemini Web To API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/v1/accounts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the status and metrics of all configured Gemini cookie accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List All Cookie Accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/providers.AccountConfig"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Injects a new __Secure-1PSID and __Secure-1PSIDTS for load balancing and rotation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Add or Update a Cookie Account",
                "parameters": [
                    {
                        "description": "Account Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AddAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/v1/accounts/clear-all-bots": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sweeps all accounts and performs bot clearance on those with errors",
                "tags": [
                    "Admin"
                ],
                "summary": "Clear All Bots",
                "responses": {}
            }
        },
        "/admin/v1/accounts/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes an account from the Rotation pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a Cookie Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/v1/accounts/{id}/clear-bot": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers the Rod browser to perform \"hãy trả lời ok\" interaction",
                "tags": [
                    "Admin"
                ],
                "summary": "Clear Bot for specific account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/v1/accounts/{id}/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Forces an immediate check or chat generation on the designated account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Test Account Rotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/v1/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the dynamic runtime configurations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get Current Configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.ConfigResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dynamically updates runtime configs like logging and chat deletion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update Configuration",
                "parameters": [
                    {
                        "description": "Configuration Toggles",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UpdateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.UpdateConfigResponse"
                        }
                    }
                }
            }
        },
        "/admin/v1/guest/discover": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers the discovery service to find and learn new AI guest platforms",
                "tags": [
                    "Guest"
                ],
                "summary": "Trigger Guest Discovery",
                "responses": {}
            }
        },
        "/admin/v1/guest/platforms": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the list of discovered and valid guest chat platforms",
                "tags": [
                    "Guest"
                ],
                "summary": "List Guest Platforms",
                "responses": {}
            }
        },
        "/admin/v1/guest/platforms/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enable or disable a discovered guest platform",
                "tags": [
                    "Guest"
                ],
                "summary": "Toggle Guest Platform Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Object: {'disabled': true}",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a discovered guest platform from the system",
                "tags": [
                    "Guest"
                ],
                "summary": "Delete Guest Platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/v1/guest/platforms/{id}/disable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually disable a discovered guest platform to skip discovery/use",
                "tags": [
                    "Guest"
                ],
                "summary": "Disable Guest Platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/v1/guest/platforms/{id}/enable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually re-enable a disabled guest platform",
                "tags": [
                    "Guest"
                ],
                "summary": "Enable Guest Platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/v1/guest/platforms/{id}/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send 'hello' to a guest platform to verify it's working",
                "tags": [
                    "Guest"
                ],
                "summary": "Test Guest Platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/v1/guest/relearn": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resets learned structure for guest platforms and initiates re-learning",
                "tags": [
                    "Guest"
                ],
                "summary": "Re-learn Guest Platforms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specific Platform ID (optional, omit for ALL)",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/admin/v1/schema/reset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resets the GJSON extraction paths to default and triggers healing",
                "tags": [
                    "Admin"
                ],
                "summary": "Reset Schema Extractor",
                "responses": {}
            }
        },
        "/admin/v1/schema/retry": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers the self-healing process to find new GJSON paths",
                "tags": [
                    "Admin"
                ],
                "summary": "Retry Healing",
                "responses": {}
            }
        },
        "/admin/v1/schema/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current GJSON schema and healing status",
                "tags": [
                    "Admin"
                ],
                "summary": "Get Schema Status",
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "admin.AddAccountRequest": {
            "type": "object",
            "properties": {
                "__Secure-1PSID": {
                    "type": "string"
                },
                "__Secure-1PSIDTS": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "admin.ConfigResponse": {
            "type": "object",
            "properties": {
                "auto_delete_chat": {
                    "type": "boolean"
                },
                "enable_guest_discovery": {
                    "type": "boolean"
                },
                "log_raw_requests": {
                    "type": "boolean"
                }
            }
        },
...
        "providers.AccountConfig": {
            "type": "object",
            "properties": {
                "__Secure-1PSID": {
                    "type": "string"
                },
                "__Secure-1PSIDTS": {
                    "type": "string"
                },
                "added_at": {
                    "type": "integer"
                },
                "error_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/providers.AccountStatus"
                },
                "success_count": {
                    "type": "integer"
                },
                "target_ip": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003cyour_admin_api_key\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}